refactor to use the latest vips client libs
