+++ src/vips-notification-worker/pom.xml
@@ -77,7 +77,7 @@
 		</dependency>
 
 		<dependency>
-			<groupId>ca.bc.gov.open</groupId>
+			<groupId>ca.bc.gov.open.pssg.rsbc</groupId>
 			<artifactId>vips-ords-client</artifactId>
 			<version>0.4.0-SNAPSHOT</version>
 		</dependency>
